backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
fastapi==0.116.2
firebase_admin==7.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.4
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
